chore: Refactoring tests. #12
