+++ test/integration/emoji_widget_test.dart
